fix lint warnings on new files
